puppet-server: Adds support for script extension.
This patch addresses the need to sometimes add arbitrary context-specific code
in a bash script. The intention in this case is to allow jobs/templates to
export environment variables specific to their needs. The global default for
this variable is the empty string which is safe for any situation.

The example of this being used is exporting the PUPPETSERVER_INSTALL_MODE
variable so that we can simply differentiate between normal install and upgrade
install presuite behavior.