(MAINT) Fix race condition in lein-deploy/ezbake pipelines.
The race condition this address is one in which `lein -U run ezbake ...` does
not acquire the latests SNAPSHOT artifact version during a run. This means, for
example, that the `future` branch pipeline of puppet-server may inadvertantly
use the locally-available SNAPSHOT without checking for a newer artifact on
nexus. If this `future` branch runs on the same host as a previous `stable`
branch ran on, the `future` pipeline then will actually be running tests against
a `stable` artifact.

Ideally, the `-U` flag alone would be enough to prevent this unexpected series
of events. However, we have found it necessary to simply remove all
locally-installed maven artifacts for the given project during the "ezbake
uberbuild" job.