(MAINT) Remove various staging projects/jobs.
I am removing these projects to be stored in my own personal branch and would
encourage everyone else to do the same.

The staging deployment type/directory is intended as a convenience for users of
ci-job-configs to experiment with various setups without deploying projects
defined in the production directory. Because of this I added empty PLACEHOLDER
files to keep these directories around and left the "staging" deployment type in
utils/deploy.sh.

It would be bad practice to push staging jobs from a development working copy to
master because:

* The staging namespace will certainly become polluted if everyone does this.
* If two people are working on ci-job-configs at the same time they may end up
  pushing conflicting changes to the same set of jobs which would lead to
  difficult-to-debug problems during pipeline development.