(maint) Filter 3.3.2/3.7.0 upgrade tests for sles12 on 3.7.x
The 3.7.x full-upgrade test is currently trying to test upgrades from PE
3.3.2 and 3.7.0 to 3.7.1 on the 3.7.x pipeline.  This doesn't work well,
because we don't have 3.3.2, 3.7.1 builds for sles12, which makes the
test angry and sad.  We do have sles12 for 3.7.1 though, so we can test
an upgrade there.  The patch attempts to create a new variable
full_upgrade_combination_filter which defaults to an empty string, and
is set in the 3.7.x instation of the pe-integration-full job-group.