Add pe-console-auth-ui
The pe-console-auth-ui is a tk-service for providing the login/logout/password reset pages and middleware for the pe-console-ui. This addition adds the project to use the "standard" simple case of clj projects (lein test / lein release).  This project differs from the other console-ui projects because the assets are static within the repo.