Make sure the js testem ports don't conflict between jobs
The testem ports used by ember-cli are currently the same for all the js jobs.  This means that if two js jobs are run at the same time on the same machine, port conflicts will cause the job to fail.  This change adds in a new testem-port variable that is used in the test script to force the job to run on a particular port.  This will prevent job port conflicts (assuming the jobs are configured correctly).