(maint) fix UI compose and build job
The UI compose and Build job has a script that references ember as an application.  This should be in the local npm bin directory.  Prior to this commit, the path to the local npm bin directory wasn't being included in the path, and thus the ember binary was unavailable.  This commit adds the local npm bin directory to the path.