Set packaging_sleep to 0 on all classifier jobs
This parameter is used when building multiple package versions off a single
source branch. If the packaging is all run at the same time it causes issues,
so we delay some. Since we aren't currently doing this, set the value to 0 but
leave the key so it's clear that it's available in the future.