(QENG-1998) Disable currently unused 4.0.x PE integration jobs
Most of the 4.0.x jobs in
http://jenkins-enterprise.delivery.puppetlabs.net/view/pe-integration/view/pe-4.0.x/
aren't going to be used by the UI and Services team (all our work is
master-side and requires no upgrade tests), who will be the primary
contributors to the 4.0.x pipeline until 3.8 goes out. However, we still care
about the smoke tests, per discussion with Britt. After talking with Wayne, it
appears that what I've supplied here is the shortest route to acheiving this -
first, we'll remove the platforms for the upgrade smoke tests, since we know
upgrades are broken right now and not getting fixed soon (known issue -
postgres upgrade). Then, under the full integration, we'll actually comment out
the jobs to remove them JJB management. Then we'll disable them in the UI. Once
4.0.x is back under active management by the integration team, we'll re-enable
these jobs.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>