(MAINT) Fix ezbake smoketests.
Before this patch we were mistakenly using the "scm_branch" value that was
intended to signify the ezbake branch we were testing on as the branch that
should be checked out on the acceptance testing repository.

This patch also adds a new job template which is specific to the ezbake
pipeline. This seems to be unavoidable due to the wonky override hierarchy in
JJB.