(maint) add release jobs for Javascript Projects
Add in an npm version script for the javascript application dependencies.  The script updates the patch version in package.json, tags the version with the new version number, and pushes the change and the tag to git.
In addition, this adds a job for an additional repo pe-base-styles.

The git includes for the js projects was incorrectly specified to dev-resources/*, which isn't appropriate for these projects.