(QENG-2071) Remove duplicate jobs, fix postgres integration job.
Previously, there were a bunch of project-specific implementations of the clj
postgres integration job. This commit removes this duplication of maintenance
effort as well as improves the postgres integration job template by always
adding the `citext` extension to the rbac databse.

However, this will leave many jobs on jenkins-enterprise in an unmanaged state;
they will need to be deleted manually.