(fixup) Remove invalid layout for classifier & rbac
In the refactor to allow periodic job kick offs for the node teams the
layouts for rbac and classifier were changed to a form of split PE
installation that is not supported at this time. This reverts these
projects to use the same layout as previous. See QENG-2036 for
commentary.